update noise analysis

+++ contribs/application/src/main/java/org/matsim/application/analysis/noise/MergeNoiseOutput.java
@@ -64,7 +64,7 @@ final class MergeNoiseOutput {
 	 *
 	 * @return maxTime value
 	 */
-	public double getMaxTime() {
+	public int getMaxTime() {
 		return maxTime;
 	}
